requirements.txt
asarPy==1.0.1
asgiref==3.8.1
beautifulsoup4==4.13.3
Django==4.2.20
django-bootstrap-v5==1.0.11
django-crispy-forms==2.3
django-debug-toolbar==5.1.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
Pygments==2.19.1
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2